job_list_service/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-decouple
django-cors-headers
requests
